Move instrument and stream validation functions to tns_robot.py for better organization.
